requirements.txt
Flask==2.0.1
google-cloud-datastore==2.1.6
json2html==1.3.0
google-auth==2.6.6
google-auth-oauthlib==0.5.1
